+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: husui <husui@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/27 16:05:41 by husui             #+#    #+#             */
-/*   Updated: 2018/01/29 00:04:19 by ztisnes          ###   ########.fr       */
+/*   Updated: 2018/01/31 21:09:59 by ztisnes          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,8 +18,8 @@
 
 char				*ft_strrev(char *str);
 char				*ft_itoa(int n);
-char				*ft_itoa2(int n);
-char				*ft_itoa_base(int n, int base);
+char				*ft_itoa2(uintmax_t n);
+char				*ft_itoa_base(uintmax_t n, int base);
 char				*ft_strcat(char *restrict s1, const char *restrict s2);
 char				*ft_strchr(const char *s, int c);
 char				*ft_strcpy(char *dst, const char *src);
